change positions for column and a row on currency cards
