+++ lib/widgets/transaction_dialog.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:seeds/constants/app_colors.dart';
 import 'package:seeds/models/models.dart';
-import 'package:seeds/screens/app/wallet/dashboard.dart';
+import 'package:seeds/screens/app/wallet/dashboard/dashboard.dart';
 import 'package:seeds/widgets/transaction_avatar.dart';
 import 'package:share/share.dart';
