+++ pubspec.yaml
@@ -282,8 +282,12 @@ flutter:
     - assets/currency/xof.png
     - assets/images/wallet/arrow_up.svg
     - assets/images/wallet/arrow_down.svg
-    - assets/images/wallet/seeds_info_card_background.jpg
-    - assets/images/wallet/seeds_info_card_logo.jpg
+    - assets/images/wallet/currency_info_cards/seeds/background.jpg
+    - assets/images/wallet/currency_info_cards/seeds/logo.jpg
+    - assets/images/wallet/currency_info_cards/hypha/background.jpg
+    - assets/images/wallet/currency_info_cards/hypha/logo.jpg
+    - assets/images/wallet/currency_info_cards/planted_seeds/logo.jpg
+    - assets/images/wallet/currency_info_cards/planted_seeds/background.jpg
     # assets end
 
   fonts:
